CallTracker: Add back-off and retry toggle to getCurrentCalls
Add a system property to handle cases where the modem
reacts slowly and does not update CURRENT_CALLS to include calls which
are dialing. A delay is introduced to avoid spamming the modem and then
GetCurrentCalls is called again in order to avoid the calltracker from
marking the call as ended when it is still active.

Tested and LS970 and E970 (Sprint and ATT LG Optimus G).

Change-Id: I7d3aac00b811bceb2909adf239492962a737de4a

Conflicts:
	src/java/com/android/internal/telephony/cdma/CdmaCallTracker.java
	src/java/com/android/internal/telephony/gsm/GsmCallTracker.java